Move static assets collector out of utils
